refactor: move RegisterAttendanceError to data.rs
