+++ src/pages/register_attendance/page.rs
@@ -1,10 +1,8 @@
-use derive_more::Display;
-
 use crate::{
     auth::is_authenticated,
-    errors::{insert_member_error::InsertMemberErr, signature_error::SignatureErr},
+    errors::signature_error::SignatureErr,
     google::sheets::insert_new_member,
-    pages::register_attendance::data::{flip_names, increment_week},
+    pages::register_attendance::data::{flip_names, increment_week, RegisterAttendanceError},
     signature::verify_signature,
     util::get_redirect_response,
     AppState,
@@ -47,40 +45,4 @@ pub async fn register_attendance(
     insert_new_member(&flipped_name, session_week_number).await?;
 
     Ok(HttpResponse::Created().body(format!("{} has been added to the roster.", &info.name)))
-}
-
-#[derive(Debug, Display)]
-pub enum RegisterAttendanceError {
-    #[display(fmt = "{}", _0)]
-    SignatureErr(SignatureErr),
-    #[display(fmt = "{}", _0)]
-    InsertMemberErr(InsertMemberErr),
-}
-
-impl actix_web::ResponseError for RegisterAttendanceError {
-    fn status_code(&self) -> actix_web::http::StatusCode {
-        match self {
-            Self::SignatureErr(s) => s.status_code(),
-            Self::InsertMemberErr(s) => s.status_code(),
-        }
-    }
-
-    fn error_response(&self) -> actix_web::HttpResponse<actix_web::body::BoxBody> {
-        match self {
-            Self::SignatureErr(s) => s.error_response(),
-            Self::InsertMemberErr(s) => s.error_response(),
-        }
-    }
-}
-
-impl From<SignatureErr> for RegisterAttendanceError {
-    fn from(value: SignatureErr) -> Self {
-        RegisterAttendanceError::SignatureErr(value)
-    }
-}
-
-impl From<InsertMemberErr> for RegisterAttendanceError {
-    fn from(value: InsertMemberErr) -> Self {
-        RegisterAttendanceError::InsertMemberErr(value)
-    }
 }
